+++ src/pnpjs/Users/mockData/getCurrentUser.ts
@@ -1,0 +1,76 @@
+
+const getFromCurrentTemplates = {
+  "user": {
+      "title": "Mike LastName",
+      "Title": "Mike LastName",
+      "initials": "MM",
+      "email": "mike.tenant@tenant.onmicrosoft.com",
+      "Email": "mike.tenant@tenant.onmicrosoft.com",
+      "LoginName": "i:0#.f|membership|mike.tenant@tenant.onmicrosoft.com",
+      "Name": "i:0#.f|membership|mike.tenant@tenant.onmicrosoft.com",
+      "id": "6",
+      "Id": "6",
+      "ID": "6",
+      "UserId": {
+          "NameId": "100320004e46552a",
+          "NameIdIssuer": "urn:federation:microsoftonline"
+      },
+      "isSiteAdmin": true,
+      "IsSiteAdmin": true,
+      "UserPrincipalName": "mike.tenant@tenant.onmicrosoft.com",
+      "imageInitials": "",
+      "imageUrl": "",
+      "loginName": "i:0#.f|membership|mike.tenant@tenant.onmicrosoft.com",
+      "text": "Mike LastName",
+      "remoteID": null,
+      "ensureWeb": "https://tenant.sharepoint.com/sites/Templates",
+      "PrincipalType": 1
+  },
+  "errorInfo": null,
+  "errorInput": {
+      "e": null,
+      "alertMe": false,
+      "consoleLog": true,
+      "traceString": "fps-library-v2|/sites/Templates/Analytics/|ErrorLog|Failed|ensureUserHere Ensure User Here ~ 24|https://tenant.sharepoint.com/sites/Templates|||",
+      "logErrors": true
+  },
+  "status": "success"
+}
+
+const getRemoteHub = {
+  "user": {
+      "title": "Mike LastName",
+      "Title": "Mike LastName",
+      "initials": "MM",
+      "email": "mike.tenant@tenant.onmicrosoft.com",
+      "Email": "mike.tenant@tenant.onmicrosoft.com",
+      "LoginName": "i:0#.f|membership|mike.tenant@tenant.onmicrosoft.com",
+      "Name": "i:0#.f|membership|mike.tenant@tenant.onmicrosoft.com",
+      "id": "9",
+      "Id": "9",
+      "ID": "9",
+      "UserId": {
+          "NameId": "100320004e46552a",
+          "NameIdIssuer": "urn:federation:microsoftonline"
+      },
+      "isSiteAdmin": true,
+      "IsSiteAdmin": true,
+      "UserPrincipalName": "mike.tenant@tenant.onmicrosoft.com",
+      "imageInitials": "",
+      "imageUrl": "",
+      "loginName": "i:0#.f|membership|mike.tenant@tenant.onmicrosoft.com",
+      "text": "Mike LastName",
+      "remoteID": null,
+      "ensureWeb": "https://tenant.sharepoint.com/sites/PivotHub",
+      "PrincipalType": 1
+  },
+  "errorInfo": null,
+  "errorInput": {
+      "e": null,
+      "alertMe": false,
+      "consoleLog": true,
+      "traceString": "fps-library-v2|/sites/Templates/Analytics/|ErrorLog|Failed|ensureUserHere Ensure User Here ~ 24|https://tenant.sharepoint.com/sites/PivotHub|||",
+      "logErrors": true
+  },
+  "status": "success"
+}