+++ src/pnpjs/Users/mockData/getUserPermissions.ts
@@ -1,0 +1,21 @@
+
+// 2022-12-10
+const onTemplatesTesting = {
+    "basePerms": {
+        "odata.metadata": "https://tenant.sharepoint.com/sites/Templates/_api/$metadata#SP.BasePermissions",
+        "High": "2147483647",
+        "Low": "4294967295"
+    },
+    "errorInfo": null,
+    "status": "success"
+}
+
+const onRemoteHub = {
+  "basePerms": {
+      "odata.metadata": "https://tenant.sharepoint.com/sites/PivotHub/_api/$metadata#SP.BasePermissions",
+      "High": "2147483647",
+      "Low": "4294705151"
+  },
+  "errorInfo": null,
+  "status": "success"
+}